Initial main.go
In this commit, we connect with the ClimaCell API using the net/http
Client from the Go standard library. We make a Request, add headers,
and then receive the response data from the http Response. Next, we
will build a client package for having a dedicated Go type handling
communication with the API.