First pass at deserializing weather data
In this commit, we define a climacell package, and based on the JSON
response of the /weather/forecast/hourly endpoint, we define types for
deserializing a weather sample and its fields.

For the most part this works. However, because fields on weather
samples in the ClimaCell API are snake_case, we need to add struct tags
to tell Go's json.Unmarshal that the observation_time field on the JSON
weather samples corresponds to the ObservationTime field on the Go
Weather struct.